+++ src/main/java/roomescape/support/exception/GlobalExceptionHandler.java
@@ -1,4 +1,4 @@
-package roomescape.support;
+package roomescape.support.exception;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
